tools/requirements.txt
PyYAML>=6.0
jsonschema>=4.0
ruamel.yaml>=0.17.21
